+++ library/src/commonTest/kotlin/dev/nesk/akkurate/accessors/IterableTest.kt
@@ -19,6 +19,7 @@ package dev.nesk.akkurate.accessors
 
 import dev.nesk.akkurate._test.Validatable
 import dev.nesk.akkurate._test.assertContentEquals
+import dev.nesk.akkurate.test.Validatable
 import dev.nesk.akkurate.validatables.Validatable
 import kotlin.test.*